Update: changing the main synchronization method and adding others
After 456b54e, a new synchronization method was added which is now called 'synchronous', however it does not work properly for all users, so the previous method was reverted as the default, which is now called 'alternately'.
An explicit choice of the 'sequential' method was also added, which is used mainly for cartesian type kinematics, has no practical application for corexy, but it can be selected.
Minor code optimization.