Update: relax belt contour after measurement
For some users, the secondary axis greatly influenced the measurements of the primary axis. For example, she reduced the amount of magnitude per step, or, on the contrary, increased it, which could lead to false measurements and confusion in the program. After the measurement, turn off the second motor on the belt contour to remove unnecessary loads. This should significantly reduce the coupling between the axes for corexy-like kinematics, but not completely, since we still have common motion paths like a beam with toolhead. Also along with this solution, the task of shifting microsteps has been moved to another stepper.