Issue: incorrect microstep scale during sync model calibration
For some configurations, the calculation of the number of microsteps per step was performed incorrectly for the calibration procedure of the sync model. Because of which the step dependence function was also found incorrectly. However, this did not cause the desired effect on the overall sync algorithm, since we are limited to stepping several microsteps per step.