Update: add priority to measured points in sync model calibration
On some printers, a sequence of measured points in the calibration is formed, to which no function can be applied correctly. Therefore, usually the functions optimal for rmse for all points are distorted near the zero point, which can lead to unexpected behavior of the calculation of microsteps during synchronization of motors. Therefore, a point priority was added, which gradually decreases from the point of 5/16 microsteps to the maximum measured points by magnitude. This will degrade the overlays quality at the highest points, but it greatly improves it at the lowest points, which is more important.