Issue: rmse calculation not on priority range in sync model calibration 92bd50b
After 92bd50b, a system of priority points, a priority range for selecting a function, was introduced. However, the rmse by which further sorting of the most suitable functions was carried out was calculated from the entire function, and not from the priority range. In some cases, especially on very curved graphs, this caused the selection of an inappropriate function.