Update: deprecate 'alternately' sync method and choice of methods
The 'synchronous' method was added almost a year ago and has proven to work better in this time. It was optional, but now it will become the main one, and the previous default method is 'alternately' and the choice of methods will no longer be available, due to its uselessness. The description of these methods has been cut from the wiki.